+++ src/init.c
@@ -152,6 +152,9 @@ extern r_obj* ffi_as_short_length(r_obj*, r_obj*);
 extern r_obj* ffi_s3_get_method(r_obj*, r_obj*, r_obj*);
 extern r_obj* ffi_list_all_size(r_obj*, r_obj*, r_obj*);
 extern r_obj* ffi_list_check_all_size(r_obj*, r_obj*, r_obj*);
+extern r_obj* ffi_vec_set_intersect(r_obj*, r_obj*, r_obj*, r_obj*);
+extern r_obj* ffi_vec_set_difference(r_obj*, r_obj*, r_obj*, r_obj*);
+extern r_obj* ffi_vec_set_union(r_obj*, r_obj*, r_obj*, r_obj*);
 
 
 // Maturing
@@ -327,6 +330,9 @@ static const R_CallMethodDef CallEntries[] = {
   {"ffi_s3_get_method",                     (DL_FUNC) &ffi_s3_get_method, 3},
   {"ffi_list_all_size",                     (DL_FUNC) &ffi_list_all_size, 3},
   {"ffi_list_check_all_size",               (DL_FUNC) &ffi_list_check_all_size, 3},
+  {"ffi_vec_set_intersect",                 (DL_FUNC) &ffi_vec_set_intersect, 4},
+  {"ffi_vec_set_difference",                (DL_FUNC) &ffi_vec_set_difference, 4},
+  {"ffi_vec_set_union",                     (DL_FUNC) &ffi_vec_set_union, 4},
   {"ffi_exp_vec_cast",                      (DL_FUNC) &exp_vec_cast, 2},
   {NULL, NULL, 0}
 };